[CALCITE-3944] Move dumpSets and dumpGraphviz out of VolcanoPlanner
