+++ core/src/main/java/org/apache/calcite/plan/volcano/VolcanoPlanner.java
@@ -16,7 +16,6 @@
  */
 package org.apache.calcite.plan.volcano;
 
-import org.apache.calcite.avatica.util.Spaces;
 import org.apache.calcite.config.CalciteConnectionConfig;
 import org.apache.calcite.config.CalciteSystemProperty;
 import org.apache.calcite.plan.AbstractRelOptPlanner;
@@ -40,7 +39,6 @@ import org.apache.calcite.plan.RelTrait;
 import org.apache.calcite.plan.RelTraitDef;
 import org.apache.calcite.plan.RelTraitSet;
 import org.apache.calcite.rel.RelNode;
-import org.apache.calcite.rel.RelVisitor;
 import org.apache.calcite.rel.convert.Converter;
 import org.apache.calcite.rel.convert.ConverterRule;
 import org.apache.calcite.rel.externalize.RelWriterImpl;
@@ -54,25 +52,20 @@ import org.apache.calcite.runtime.Hook;
 import org.apache.calcite.sql.SqlExplainLevel;
 import org.apache.calcite.util.Litmus;
 import org.apache.calcite.util.Pair;
-import org.apache.calcite.util.PartiallyOrderedSet;
 import org.apache.calcite.util.Util;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.LinkedListMultimap;
 import com.google.common.collect.Multimap;
-import com.google.common.collect.Ordering;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.util.ArrayDeque;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Comparator;
 import java.util.Deque;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.IdentityHashMap;
-import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -531,7 +524,7 @@ public class VolcanoPlanner extends AbstractRelOptPlanner {
           "Cheapest plan:\n{}", RelOptUtil.toString(cheapest, SqlExplainLevel.ALL_ATTRIBUTES));
 
       if (!provenanceMap.isEmpty()) {
-        LOGGER.debug("Provenance:\n{}", provenance(cheapest));
+        LOGGER.debug("Provenance:\n{}", Dumpers.provenance(provenanceMap, cheapest));
       }
     }
     return cheapest;
@@ -574,74 +567,6 @@ public class VolcanoPlanner extends AbstractRelOptPlanner {
                 difference.get(0).getTraitDef(), root.getTraitSet()),
             root);
       }
-    }
-  }
-
-  /**
-   * Returns a multi-line string describing the provenance of a tree of
-   * relational expressions. For each node in the tree, prints the rule that
-   * created the node, if any. Recursively describes the provenance of the
-   * relational expressions that are the arguments to that rule.
-   *
-   * <p>Thus, every relational expression and rule invocation that affected
-   * the final outcome is described in the provenance. This can be useful
-   * when finding the root cause of "mistakes" in a query plan.</p>
-   *
-   * @param root Root relational expression in a tree
-   * @return Multi-line string describing the rules that created the tree
-   */
-  private String provenance(RelNode root) {
-    final StringWriter sw = new StringWriter();
-    final PrintWriter pw = new PrintWriter(sw);
-    final List<RelNode> nodes = new ArrayList<>();
-    new RelVisitor() {
-      public void visit(RelNode node, int ordinal, RelNode parent) {
-        nodes.add(node);
-        super.visit(node, ordinal, parent);
-      }
-      // CHECKSTYLE: IGNORE 1
-    }.go(root);
-    final Set<RelNode> visited = new HashSet<>();
-    for (RelNode node : nodes) {
-      provenanceRecurse(pw, node, 0, visited);
-    }
-    pw.flush();
-    return sw.toString();
-  }
-
-  /**
-   * Helper for {@link #provenance(org.apache.calcite.rel.RelNode)}.
-   */
-  private void provenanceRecurse(
-      PrintWriter pw, RelNode node, int i, Set<RelNode> visited) {
-    Spaces.append(pw, i * 2);
-    if (!visited.add(node)) {
-      pw.println("rel#" + node.getId() + " (see above)");
-      return;
-    }
-    pw.println(node);
-    final Provenance o = provenanceMap.get(node);
-    Spaces.append(pw, i * 2 + 2);
-    if (o == Provenance.EMPTY) {
-      pw.println("no parent");
-    } else if (o instanceof DirectProvenance) {
-      RelNode rel = ((DirectProvenance) o).source;
-      pw.println("direct");
-      provenanceRecurse(pw, rel, i + 2, visited);
-    } else if (o instanceof RuleProvenance) {
-      RuleProvenance rule = (RuleProvenance) o;
-      pw.println("call#" + rule.callId + " rule [" + rule.rule + "]");
-      for (RelNode rel : rule.rels) {
-        provenanceRecurse(pw, rel, i + 2, visited);
-      }
-    } else if (o == null && node instanceof RelSubset) {
-      // A few operands recognize subsets, not individual rels.
-      // The first rel in the subset is deemed to have created it.
-      final RelSubset subset = (RelSubset) node;
-      pw.println("subset " + subset);
-      provenanceRecurse(pw, subset.getRelList().get(0), i + 2, visited);
-    } else {
-      throw new AssertionError("bad type " + o);
     }
   }
 
@@ -912,12 +837,12 @@ public class VolcanoPlanner extends AbstractRelOptPlanner {
       if (CalciteSystemProperty.DUMP_SETS.value()) {
         pw.println();
         pw.println("Sets:");
-        dumpSets(pw);
+        Dumpers.dumpSets(this, pw);
       }
       if (CalciteSystemProperty.DUMP_GRAPHVIZ.value()) {
         pw.println();
         pw.println("Graphviz:");
-        dumpGraphviz(pw);
+        Dumpers.dumpGraphviz(this, pw);
       }
     } catch (Exception | AssertionError e) {
       pw.println("Error when dumping plan state: \n"
@@ -928,229 +853,9 @@ public class VolcanoPlanner extends AbstractRelOptPlanner {
   public String toDot() {
     StringWriter sw = new StringWriter();
     PrintWriter pw = new PrintWriter(sw);
-    dumpGraphviz(pw);
+    Dumpers.dumpGraphviz(this, pw);
     pw.flush();
     return sw.toString();
-  }
-
-  private void dumpSets(PrintWriter pw) {
-    Ordering<RelSet> ordering = Ordering.from(Comparator.comparingInt(o -> o.id));
-    for (RelSet set : ordering.immutableSortedCopy(allSets)) {
-      pw.println("Set#" + set.id
-          + ", type: " + set.subsets.get(0).getRowType());
-      int j = -1;
-      for (RelSubset subset : set.subsets) {
-        ++j;
-        pw.println(
-            "\t" + subset + ", best="
-            + ((subset.best == null) ? "null"
-                : ("rel#" + subset.best.getId())));
-        assert subset.set == set;
-        for (int k = 0; k < j; k++) {
-          assert !set.subsets.get(k).getTraitSet().equals(
-              subset.getTraitSet());
-        }
-        for (RelNode rel : subset.getRels()) {
-          // "\t\trel#34:JavaProject(rel#32:JavaFilter(...), ...)"
-          pw.print("\t\t" + rel);
-          for (RelNode input : rel.getInputs()) {
-            RelSubset inputSubset =
-                getSubset(
-                    input,
-                    input.getTraitSet());
-            RelSet inputSet = inputSubset.set;
-            if (input instanceof RelSubset) {
-              final Iterator<RelNode> rels =
-                  inputSubset.getRels().iterator();
-              if (rels.hasNext()) {
-                input = rels.next();
-                assert input.getTraitSet().satisfies(inputSubset.getTraitSet());
-                assert inputSet.rels.contains(input);
-                assert inputSet.subsets.contains(inputSubset);
-              }
-            }
-          }
-          if (prunedNodes.contains(rel)) {
-            pw.print(", pruned");
-          }
-          RelMetadataQuery mq = rel.getCluster().getMetadataQuery();
-          pw.print(", rowcount=" + mq.getRowCount(rel));
-          pw.println(", cumulative cost=" + getCost(rel, mq));
-        }
-      }
-    }
-  }
-
-  private void dumpGraphviz(PrintWriter pw) {
-    Ordering<RelSet> ordering = Ordering.from(Comparator.comparingInt(o -> o.id));
-    Set<RelNode> activeRels = new HashSet<>();
-    for (VolcanoRuleCall volcanoRuleCall : ruleCallStack) {
-      activeRels.addAll(Arrays.asList(volcanoRuleCall.rels));
-    }
-    pw.println("digraph G {");
-    pw.println("\troot [style=filled,label=\"Root\"];");
-    PartiallyOrderedSet<RelSubset> subsetPoset = new PartiallyOrderedSet<>(
-        (e1, e2) -> e1.getTraitSet().satisfies(e2.getTraitSet()));
-    Set<RelSubset> nonEmptySubsets = new HashSet<>();
-    for (RelSet set : ordering.immutableSortedCopy(allSets)) {
-      pw.print("\tsubgraph cluster");
-      pw.print(set.id);
-      pw.println("{");
-      pw.print("\t\tlabel=");
-      Util.printJavaString(pw, "Set " + set.id + " " + set.subsets.get(0).getRowType(), false);
-      pw.print(";\n");
-      for (RelNode rel : set.rels) {
-        pw.print("\t\trel");
-        pw.print(rel.getId());
-        pw.print(" [label=");
-        RelMetadataQuery mq = rel.getCluster().getMetadataQuery();
-
-        // Note: rel traitset could be different from its subset.traitset
-        // It can happen due to RelTraitset#simplify
-        // If the traits are different, we want to keep them on a graph
-        String traits = "." + getSubset(rel).getTraitSet().toString();
-        String title = rel.toString().replace(traits, "");
-        if (title.endsWith(")")) {
-          int openParen = title.indexOf('(');
-          if (openParen != -1) {
-            // Title is like rel#12:LogicalJoin(left=RelSubset#4,right=RelSubset#3,
-            // condition==($2, $0),joinType=inner)
-            // so we remove the parenthesis, and wrap parameters to the second line
-            // This avoids "too wide" Graphiz boxes, and makes the graph easier to follow
-            title = title.substring(0, openParen) + '\n'
-                + title.substring(openParen + 1, title.length() - 1);
-          }
-        }
-        Util.printJavaString(pw,
-            title
-                + "\nrows=" + mq.getRowCount(rel) + ", cost=" + getCost(rel, mq), false);
-        RelSubset relSubset = getSubset(rel);
-        if (!(rel instanceof AbstractConverter)) {
-          nonEmptySubsets.add(relSubset);
-        }
-        if (relSubset.best == rel) {
-          pw.print(",color=blue");
-        }
-        if (activeRels.contains(rel)) {
-          pw.print(",style=dashed");
-        }
-        pw.print(",shape=box");
-        pw.println("]");
-      }
-
-      subsetPoset.clear();
-      for (RelSubset subset : set.subsets) {
-        subsetPoset.add(subset);
-        pw.print("\t\tsubset");
-        pw.print(subset.getId());
-        pw.print(" [label=");
-        Util.printJavaString(pw, subset.toString(), false);
-        boolean empty = !nonEmptySubsets.contains(subset);
-        if (empty) {
-          // We don't want to iterate over rels when we know the set is not empty
-          for (RelNode rel : subset.getRels()) {
-            if (!(rel instanceof AbstractConverter)) {
-              empty = false;
-              break;
-            }
-          }
-          if (empty) {
-            pw.print(",color=red");
-          }
-        }
-        if (activeRels.contains(subset)) {
-          pw.print(",style=dashed");
-        }
-        pw.print("]\n");
-      }
-
-      for (RelSubset subset : subsetPoset) {
-        for (RelSubset parent : subsetPoset.getChildren(subset)) {
-          pw.print("\t\tsubset");
-          pw.print(subset.getId());
-          pw.print(" -> subset");
-          pw.print(parent.getId());
-          pw.print(";");
-        }
-      }
-
-      pw.print("\t}\n");
-    }
-    // Note: it is important that all the links are declared AFTER declaration of the nodes
-    // Otherwise Graphviz creates nodes implicitly, and puts them into a wrong cluster
-    pw.print("\troot -> subset");
-    pw.print(root.getId());
-    pw.println(";");
-    for (RelSet set : ordering.immutableSortedCopy(allSets)) {
-      for (RelNode rel : set.rels) {
-        RelSubset relSubset = getSubset(rel);
-        pw.print("\tsubset");
-        pw.print(relSubset.getId());
-        pw.print(" -> rel");
-        pw.print(rel.getId());
-        if (relSubset.best == rel) {
-          pw.print("[color=blue]");
-        }
-        pw.print(";");
-        List<RelNode> inputs = rel.getInputs();
-        for (int i = 0; i < inputs.size(); i++) {
-          RelNode input = inputs.get(i);
-          pw.print(" rel");
-          pw.print(rel.getId());
-          pw.print(" -> ");
-          pw.print(input instanceof RelSubset ? "subset" : "rel");
-          pw.print(input.getId());
-          if (relSubset.best == rel || inputs.size() > 1) {
-            char sep = '[';
-            if (relSubset.best == rel) {
-              pw.print(sep);
-              pw.print("color=blue");
-              sep = ',';
-            }
-            if (inputs.size() > 1) {
-              pw.print(sep);
-              pw.print("label=\"");
-              pw.print(i);
-              pw.print("\"");
-              // sep = ',';
-            }
-            pw.print(']');
-          }
-          pw.print(";");
-        }
-        pw.println();
-      }
-    }
-
-    // Draw lines for current rules
-    for (VolcanoRuleCall ruleCall : ruleCallStack) {
-      pw.print("rule");
-      pw.print(ruleCall.id);
-      pw.print(" [style=dashed,label=");
-      Util.printJavaString(pw, ruleCall.rule.toString(), false);
-      pw.print("]");
-
-      RelNode[] rels = ruleCall.rels;
-      for (int i = 0; i < rels.length; i++) {
-        RelNode rel = rels[i];
-        pw.print(" rule");
-        pw.print(ruleCall.id);
-        pw.print(" -> ");
-        pw.print(rel instanceof RelSubset ? "subset" : "rel");
-        pw.print(rel.getId());
-        pw.print(" [style=dashed");
-        if (rels.length > 1) {
-          pw.print(",label=\"");
-          pw.print(i);
-          pw.print("\"");
-        }
-        pw.print("]");
-        pw.print(";");
-      }
-      pw.println();
-    }
-
-    pw.print("}");
   }
 
   /**
@@ -1692,7 +1397,7 @@ public class VolcanoPlanner extends AbstractRelOptPlanner {
   /**
    * Where a RelNode came from.
    */
-  private abstract static class Provenance {
+  abstract static class Provenance {
     public static final Provenance EMPTY = new UnknownProvenance();
   }
